requirements.txt
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
dj-database-url==2.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
exceptiongroup==1.1.2
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
interruptingcow==0.8
lyricsgenius==3.0.1
moviepy==1.0.3
numpy==1.25.2
outcome==1.2.0
Pillow==9.5.0
proglog==0.1.10
psycopg2==2.9.7
PySocks==1.7.1
python-decouple==3.8
pytube==15.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
retrying==1.3.4
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spotipy==2.23.0
sqlparse==0.4.4
syrics==0.0.1.8
tinytag==1.9.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
wsproto==1.2.0
youtube-search-python==1.6.6
